src/requirements.txt
langchain==0.2.12
langchain-community==0.2.11
langchain-openai==0.1.20
pydantic-settings==2.4.0
fastapi==0.112.0
uvicorn==0.30.5
bs4==0.0.2
PyYAML
